Rename(Item): "Item 폴더 구조 변경"
Item 폴더 구조를 더 보기 쉽게 바꾸고 불필요해진 gitkeep 파일을 삭제하였습니다.
