--- src/main/java/com/space/server/core/domain/item/value/Category.java
@@ -1,6 +1,0 @@
-package com.space.server.core.domain.item.value;
-
-public enum Category {
-  HEAD,
-  THEME
-}
