+++ src/main/java/com/space/server/core/item/domain/value/Category.java
@@ -1,0 +1,6 @@
+package com.space.server.core.item.domain.value;
+
+public enum Category {
+  HEAD,
+  THEME
+}
